requirements.txt
streamlit==1.38.0
pillow==10.4.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
requests==2.32.3
ImageHash==4.3.1
